refactor: remove usdc-paymaster
