+++ src/paymasters/index.ts
@@ -1,2 +1,1 @@
-export * from './usdc-paymaster'
 export * from './public-paymaster'
