+++ test/scripts/execute-kernel-usdc-paymaster.ts
@@ -3,7 +3,7 @@ import { ADDRESS } from '@/addresses'
 import { ERC4337Bundler, type TypedData } from '@/core'
 import { fetchGasPricePimlico } from '@/index'
 import { INTERFACES } from '@/interfaces'
-import { createUSDCPaymaster } from '@/paymasters/usdc-paymaster'
+import { createUSDCPaymaster } from 'test/usdc-paymaster'
 import { toBytes32 } from '@/utils'
 import { getECDSAValidator } from '@/validations/getECDSAValidator'
 import { SingleEOAValidation } from '@/validations/SingleEOAValidation'
