+++ test/scripts/execute-simple7702-usdc-paymaster.ts
@@ -3,7 +3,7 @@ import { ADDRESS } from '@/addresses'
 import { ERC4337Bundler, type TypedData } from '@/core'
 import { fetchGasPricePimlico } from '@/fetchGasPrice'
 import { INTERFACES } from '@/interfaces'
-import { createUSDCPaymaster } from '@/paymasters/usdc-paymaster'
+import { createUSDCPaymaster } from 'test/usdc-paymaster'
 import { JsonRpcProvider, parseUnits, Wallet } from 'ethers'
 import { alchemy, pimlico } from 'evm-providers'
 import { executeUserOperation } from '../helpers'
